Create and sketch the symbol table structure
Not a bad first cut at a symbol table...
Symbols will be keyed in the global lookup table using the Clojure symbol
pathing mod the final slash. This allows for the *ns* value to operate as
a stack of namespaces in order from deepest to shallowest, and makes symbol
resolution a simple exercise in searching & popping.

TODO:
 - Add functions descend! and rise! for going up and down the ns stack
 - Add a search function for searching the current namespace and parents
   via recursion over the (pop) of the stack.
 - Figure out how to deal with tracking labels
 - Add the install! function, which adds a symbol to the current level
 - Modify the grammar with semantics hooks so that programs and functions
   correctly effect the namespace stack.
 - Write up a spec for the structure of record entries
 - Figure out how to resolve the collision case of types and functions
   sharing the same name value in the same namespace... add an error in
   the install! routine?