tweak to provide an identity convention on types
For records used only to represent types rather than instances of types, I
define `(type this)` to be `this`. This allows me to make a (typeof) call
standard on any type argument because for primitive types it won't matter,
whereas for ThinType instances the call will pass through all the way down
to the base primitive or record type at the bottom.