rework array and string naming
Strings are now installed as genvars with array<alpha>[len] types.
Arrays are now named as "array<alpha>[len]" in the symbol table.
These two changes are the stepping stones to allowing for assignments
between symbols constituting arrays of equal length. Note that this
behavior as with assigment between record types is only supported by the
expression type checker, not yet by the code generator.